clean dataset code for caddn model
