+++ configs/caddn/caddn_resnet101_kitti_iter74240.yaml
@@ -3,7 +3,7 @@ iters: 74240  # 928*80
 sync_bn: true
 
 train_dataset:
-  type: KittiCadnnDataset
+  type: KittiDepthDataset
   dataset_root: data/kitti
   point_cloud_range: [2, -30.08, -3.0, 46.8, 30.08, 1.0]
   depth_downsample_factor: 4
@@ -12,7 +12,7 @@ train_dataset:
   mode: train
 
 val_dataset:
-  type: KittiCadnnDataset
+  type: KittiDepthDataset
   dataset_root: data/kitti
   point_cloud_range: [2, -30.08, -3.0, 46.8, 30.08, 1.0]
   depth_downsample_factor: 4
