--- paddle3d/datasets/cadnn_kitti/data_processor.py
@@ -1,84 +1,0 @@
-# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from functools import partial
-
-import numpy as np
-
-from ...utils import box_utils, common_utils
-
-
-class DataProcessor(object):
-    """
-    This function refers to https://github.com/TRAILab/CaDDN/blob/5a96b37f16b3c29dd2509507b1cdfdff5d53c558/pcdet/datasets/processor/data_processor.py#L13
-    """
-
-    def __init__(self, processor_configs, point_cloud_range, training):
-        self.point_cloud_range = point_cloud_range
-        self.training = training
-        self.mode = 'train' if training else 'test'
-        self.grid_size = self.voxel_size = None
-
-        self.data_processor_queue = []
-        for cur_cfg in processor_configs:
-            cur_processor = getattr(self, cur_cfg.NAME)(config=cur_cfg)
-            self.data_processor_queue.append(cur_processor)
-
-    def mask_points_and_boxes_outside_range(self, data_dict=None, config=None):
-        if data_dict is None:
-            return partial(
-                self.mask_points_and_boxes_outside_range, config=config)
-
-        if data_dict.get('points', None) is not None:
-            mask = common_utils.mask_points_by_range(data_dict['points'],
-                                                     self.point_cloud_range)
-            data_dict['points'] = data_dict['points'][mask]
-
-        if data_dict.get(
-                'gt_boxes', None
-        ) is not None and config.REMOVE_OUTSIDE_BOXES and self.training:
-            mask = box_utils.mask_boxes_outside_range_numpy(
-                data_dict['gt_boxes'],
-                self.point_cloud_range,
-                min_num_corners=config.get('min_num_corners', 1))
-            data_dict['gt_boxes'] = data_dict['gt_boxes'][mask]
-        return data_dict
-
-    def calculate_grid_size(self, data_dict=None, config=None):
-        if data_dict is None:
-            grid_size = (self.point_cloud_range[3:6] -
-                         self.point_cloud_range[0:3]) / np.array(
-                             config.VOXEL_SIZE)
-            self.grid_size = np.round(grid_size).astype(np.int64)
-            self.voxel_size = config.VOXEL_SIZE
-            return partial(self.calculate_grid_size, config=config)
-        else:
-            return data_dict
-
-    def forward(self, data_dict):
-        """
-        Args:
-            data_dict:
-                points: (N, 3 + C_in)
-                gt_boxes: optional, (N, 7 + C) [x, y, z, dx, dy, dz, heading, ...]
-                gt_names: optional, (N), string
-                ...
-
-        Returns:
-        """
-
-        for cur_processor in self.data_processor_queue:
-            data_dict = cur_processor(data_dict=data_dict)
-
-        return data_dict
