+++ paddle3d/datasets/kitti/__init__.py
@@ -14,3 +14,4 @@
 
 from .kitti_mono_det import KittiMonoDataset
 from .kitti_pointcloud_det import KittiPCDataset
+from .kitti_depth_det import KittiDepthDataset
