+++ paddle3d/transforms/__init__.py
@@ -17,3 +17,4 @@ from paddle3d.transforms.batch_operators import *
 from paddle3d.transforms.detection import *
 from paddle3d.transforms.normalize import *
 from paddle3d.transforms.reader import *
+from paddle3d.transforms.functional import *
