+++ paddle3d/transforms/functional.py
@@ -15,6 +15,7 @@
 from typing import Tuple
 
 import numba
+import copy
 import numpy as np
 
 from paddle3d.geometries.bbox import (box_collision_test, iou_2d_jit,
@@ -287,3 +288,55 @@ def nearest_iou_similarity(bboxes_3d_1, bboxes_3d_2):
     boxes_bv_1 = rbbox2d_to_near_bbox(bboxes_3d_1[:, [0, 1, 3, 4, 6]])
     boxes_bv_2 = rbbox2d_to_near_bbox(bboxes_3d_2[:, [0, 1, 3, 4, 6]])
     return iou_2d_jit(boxes_bv_1, boxes_bv_2)
+
+def random_depth_image_horizontal(data_dict=None):
+    """
+    Performs random horizontal flip augmentation
+    Args:
+    data_dict:
+        image [np.ndarray(H_image, W_image, 3)]: Image
+        depth_map [np.ndarray(H_depth, W_depth]: Depth map
+        gt_boxes [np.ndarray(N, 7)]: 3D box labels in LiDAR coordinates [x, y, z, w, l, h, ry]
+        calib [calibration.Calibration]: Calibration object
+    Returns:
+    data_dict:
+        aug_image [np.ndarray(H_image, W_image, 3)]: Augmented image
+        aug_depth_map [np.ndarray(H_depth, W_depth]: Augmented depth map
+        aug_gt_boxes [np.ndarray(N, 7)]: Augmented 3D box labels in LiDAR coordinates [x, y, z, w, l, h, ry]
+    """
+    if data_dict is None:
+        return
+    image = data_dict["images"]
+    depth_map = data_dict["depth_maps"]
+    gt_boxes = data_dict['gt_boxes']
+    calib = data_dict["calib"]
+
+    # Randomly augment with 50% chance
+    enable = np.random.choice([False, True], replace=False, p=[0.5, 0.5])
+
+    if enable:
+        # Flip images
+        aug_image = np.fliplr(image)
+        aug_depth_map = np.fliplr(depth_map)
+
+        # Flip 3D gt_boxes by flipping the centroids in image space
+        aug_gt_boxes = copy.copy(gt_boxes)
+        locations = aug_gt_boxes[:, :3]
+        img_pts, img_depth = calib.lidar_to_img(locations)
+        W = image.shape[1]
+        img_pts[:, 0] = W - img_pts[:, 0]
+        pts_rect = calib.img_to_rect(
+            u=img_pts[:, 0], v=img_pts[:, 1], depth_rect=img_depth)
+        pts_lidar = calib.rect_to_lidar(pts_rect)
+        aug_gt_boxes[:, :3] = pts_lidar
+        aug_gt_boxes[:, 6] = -1 * aug_gt_boxes[:, 6]
+    else:
+        aug_image = image
+        aug_depth_map = depth_map
+        aug_gt_boxes = gt_boxes
+    
+    data_dict['images'] = aug_image
+    data_dict['depth_maps'] = aug_depth_map
+    data_dict['gt_boxes'] = aug_gt_boxes
+
+    return data_dict
